Removes CacheChecker.
