+++ jib-core/src/main/java/com/google/cloud/tools/jib/cache/CacheReader.java
@@ -16,19 +16,68 @@
 
 package com.google.cloud.tools.jib.cache;
 
+import com.google.cloud.tools.jib.image.DescriptorDigest;
 import com.google.cloud.tools.jib.image.ImageLayers;
+import com.google.cloud.tools.jib.image.LayerPropertyNotFoundException;
+import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
 import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
 
 /** Reads image content from the cache. */
 public class CacheReader {
+
+  /**
+   * @return the last modified time for the file at {@code path}. Recursively finds the most recent
+   *     last modified time for all subfiles if the file is a directory.
+   */
+  private static FileTime getLastModifiedTime(Path path) throws IOException {
+    FileTime lastModifiedTime = Files.getLastModifiedTime(path);
+
+    if (Files.isReadable(path)) {
+      try {
+        Optional<FileTime> maxLastModifiedTime =
+            Files.walk(path)
+                .map(
+                    subFilePath -> {
+                      try {
+                        return Files.getLastModifiedTime(subFilePath);
+
+                      } catch (IOException ex) {
+                        throw new UncheckedIOException(ex);
+                      }
+                    })
+                .max(FileTime::compareTo);
+
+        if (!maxLastModifiedTime.isPresent()) {
+          throw new IllegalStateException(
+              "Could not get last modified time for all files in directory '" + path + "'");
+        }
+        if (maxLastModifiedTime.get().compareTo(lastModifiedTime) > 0) {
+          lastModifiedTime = maxLastModifiedTime.get();
+        }
+
+      } catch (UncheckedIOException ex) {
+        throw ex.getCause();
+      }
+    }
+
+    return lastModifiedTime;
+  }
 
   private final Cache cache;
 
   public CacheReader(Cache cache) {
     this.cache = cache;
+  }
+
+  /** @return the cached layer with digest {@code layerDigest} */
+  public CachedLayer getLayer(DescriptorDigest layerDigest) throws LayerPropertyNotFoundException {
+    return cache.getMetadata().getLayers().get(layerDigest);
   }
 
   /**
@@ -65,5 +114,42 @@ public class CacheReader {
       default:
         throw new UnsupportedOperationException("Can only find layer files for application layers");
     }
+  }
+
+  /**
+   * Gets an up-to-date layer that is built from the {@code sourceFiles}.
+   *
+   * <p>The method returns the first up-to-date layer found. This is safe because the source files
+   * will not have been modified since creation of any up-to-date layer (ie. all up-to-date layers
+   * should have the same file contents)
+   */
+  public CachedLayer getUpToDateLayerBySourceFiles(List<Path> sourceFiles)
+      throws IOException, CacheMetadataCorruptedException {
+    // Grabs all the layers that have matching source files.
+    ImageLayers<CachedLayerWithMetadata> cachedLayersWithSourceFiles =
+        cache.getMetadata().filterLayers().bySourceFiles(sourceFiles).filter();
+    if (cachedLayersWithSourceFiles.isEmpty()) {
+      return null;
+    }
+
+    // Determines the latest modification time for the source files.
+    FileTime sourceFilesLastModifiedTime = FileTime.from(Instant.MIN);
+    for (Path path : sourceFiles) {
+      FileTime lastModifiedTime = getLastModifiedTime(path);
+      if (lastModifiedTime.compareTo(sourceFilesLastModifiedTime) > 0) {
+        sourceFilesLastModifiedTime = lastModifiedTime;
+      }
+    }
+
+    // Checks if at least one of the matched layers is up-to-date.
+    for (CachedLayerWithMetadata cachedLayer : cachedLayersWithSourceFiles) {
+      if (sourceFilesLastModifiedTime.compareTo(cachedLayer.getMetadata().getLastModifiedTime())
+          <= 0) {
+        // This layer is an up-to-date layer.
+        return cachedLayer;
+      }
+    }
+
+    return null;
   }
 }
